# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"AI Powered Form Filling API","version":"1.0.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login For Access Token","operationId":"login_for_access_token_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Authentication"],"summary":"Create User","operationId":"create_user_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Read Users Me","operationId":"read_users_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/templates/template":{"get":{"tags":["Templates"],"summary":"Get Templates","operationId":"get_templates_api_v1_templates_template_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Templates"],"summary":"Create Template","description":"Upload a new HTML template.\nAccepts .html files for flexible form filling.","operationId":"create_template_api_v1_templates_template_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_template_api_v1_templates_template_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Templates"],"summary":"Update Template","description":"Update an existing HTML template.","operationId":"update_template_api_v1_templates_template_put","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_update_template_api_v1_templates_template_put"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Templates"],"summary":"Delete Template","description":"Delete a template and its file.","operationId":"delete_template_api_v1_templates_template_delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"template_id","in":"query","required":true,"schema":{"type":"integer","title":"Template Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/templates/template/{template_id}/preview":{"get":{"tags":["Templates"],"summary":"Preview Template","description":"Get template HTML for preview.","operationId":"preview_template_api_v1_templates_template__template_id__preview_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"template_id","in":"path","required":true,"schema":{"type":"integer","title":"Template Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/form-fill/":{"post":{"tags":["Form Fill"],"summary":"Form Fill","description":"Fill HTML template with extracted entity data.\nReturns filled HTML content.","operationId":"form_fill_api_v1_form_fill__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"template_id","in":"query","required":true,"schema":{"type":"integer","title":"Template Id"}},{"name":"entity_id","in":"query","required":true,"schema":{"type":"integer","title":"Entity Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/form-fill/{user_id}/{filename}":{"get":{"tags":["Form Fill"],"summary":"Get Filled Form","description":"Retrieve a filled HTML form file.","operationId":"get_filled_form_api_v1_form_fill__user_id___filename__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/form-fill/preview/{template_id}/{entity_id}":{"get":{"tags":["Form Fill"],"summary":"Preview Filled Form","description":"Preview filled form as HTML response.","operationId":"preview_filled_form_api_v1_form_fill_preview__template_id___entity_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"template_id","in":"path","required":true,"schema":{"type":"integer","title":"Template Id"}},{"name":"entity_id","in":"path","required":true,"schema":{"type":"integer","title":"Entity Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/entities/entities/":{"get":{"tags":["entities"],"summary":"List Entities","description":"List entities with pagination.","operationId":"list_entities_api_v1_entities_entities__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["entities"],"summary":"Create Entity","description":"Create a new entity.","operationId":"create_entity_api_v1_entities_entities__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"entity_name","in":"query","required":true,"schema":{"type":"string","title":"Entity Name"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/entities/entities/{entity_id}":{"get":{"tags":["entities"],"summary":"Get Entity","description":"Get a single entity by id.","operationId":"get_entity_api_v1_entities_entities__entity_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"entity_id","in":"path","required":true,"schema":{"type":"integer","title":"Entity Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["entities"],"summary":"Update Entity","description":"Update an existing entity.","operationId":"update_entity_api_v1_entities_entities__entity_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"entity_id","in":"path","required":true,"schema":{"type":"integer","title":"Entity Id"}},{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["entities"],"summary":"Delete Entity","description":"Delete an entity.","operationId":"delete_entity_api_v1_entities_entities__entity_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"entity_id","in":"path","required":true,"schema":{"type":"integer","title":"Entity Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/entities-data/entities-data/":{"get":{"tags":["entities-data"],"summary":"List Entity Data","description":"Get consolidated data for a specific entity.","operationId":"list_entity_data_api_v1_entities_data_entities_data__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"entity_id","in":"query","required":true,"schema":{"type":"integer","title":"Entity Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["entities-data"],"summary":"Create Entity Data","description":"Create or update extracted data record for an entity.\nData is merged into a single consolidated record per entity.","operationId":"create_entity_data_api_v1_entities_data_entities_data__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"entity_id","in":"query","required":true,"schema":{"type":"integer","title":"Entity Id"}},{"name":"lang","in":"query","required":false,"schema":{"type":"string","default":"en","title":"Lang"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_entity_data_api_v1_entities_data_entities_data__post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_create_entity_data_api_v1_entities_data_entities_data__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_create_entity_data_api_v1_entities_data_entities_data__post"},"Body_create_template_api_v1_templates_template_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"lang":{"type":"string","title":"Lang"}},"type":"object","required":["file"],"title":"Body_create_template_api_v1_templates_template_post"},"Body_update_template_api_v1_templates_template_put":{"properties":{"template_id":{"type":"integer","title":"Template Id"},"file":{"type":"string","format":"binary","title":"File"},"lang":{"type":"string","title":"Lang"}},"type":"object","required":["template_id"],"title":"Body_update_template_api_v1_templates_template_put"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"}},"type":"object","required":["id","username"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}